+++ src/NonPerishableItem.cpp
@@ -3,7 +3,7 @@
 /* 
  * Constructor for NonPerishableItem; call's base class constructor.
 */
-NonPerishableItem::NonPerishableItem(const std::string& nm, const std::string& cat, unsigned int qty, unsigned int idd, double price, double cost, double tx)
+NonPerishableItem::NonPerishableItem(const std::string& nm, const std::string& cat, unsigned long qty, unsigned long idd, double price, double cost, double tx)
 : Item(nm, cat, qty, idd, price, cost, tx)
 {
